fix Tags and Categories pages
The idea was to have Tags and Categories as normal pages and not part of
the sidebar since they don't offer much usefulness for my blog. However,
when you go to the Tabs or Categories page from the breadcumbs when viewing
a specific tag/category, the URL will have a trailing slash, which GitHub
pages doesn't handle apparantly. So, I am including them back in the
sidebar to have the pages working correctly again.

Note that the local dev server of Jekyll (or whoever provides it when
running `bundle exec jekyll s`) does handle the trailing slash.