fix: don't use sticky position for Vita header
I thought that was a cool effect but the items in the vita are to short
too make really good use of that look. Also, on mobile this is kinda
broken. I guess maybe due to the line breaks... I don't know, I didn't
investigate it deeply, since the sticky effect wasn't that useful in the
first place