style: improve title of Postcard for mobile layouts
Previously, the title clamped after 1 line. This is problematic on
mobile layouts since there are only a few words visible. Increasing
the clamping is gonna be difficult since the positioning needs to be
dynamic depending on the number of lines... So simplicity instead just
put the title in the PostCard on mobile layouts.